+++ Code/Client/users/src/App.js
@@ -9,7 +9,7 @@ import Career from './pages/careers/Career';
 import Services from './pages/Services';
 import ErrorPage from './pages/unknownPage/ErrorPage';
 import LandingCarPage from './pages/carRental/LandingCarPage';
-import ExtendedDetailPage from './pages/carRental/ExtendedDetailPage';
+import ExtendedDetailPage from './pages/carRental/addOnPage/ExtendedDetailPage';
 import RentalBookingPage from './pages/carRental/bookingPage/RentalBookingPage'
 
 function App() {
