+++ Code/Client/users/src/components/atoms/CarRental/AddOnsCard.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function AddOnsCard() {
+  return (
+    <div>AddOnsCard</div>
+  )
+}
