+++ Code/Client/users/src/components/atoms/CarRental/RentalPriceCard.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function RentalPriceCard() {
+  return (
+    <div>RentalPriceCard</div>
+  )
+}
