+++ Code/Client/users/src/pages/CarRental/addOnsPage/ExtendedDetail.module.css
@@ -1,0 +1,3 @@
+.container{
+    background-color: rgba(235, 235, 240, 1)
+}
